% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ase}
\alias{ase}
\title{Adjacency Spectral Embedding (ASE)}
\usage{
ase(M,d)
}
\arguments{
\item{M}{A symmetric and positive semi-definite matrix.}

\item{d}{A non-negative integer embedding dimension.}
}
\value{
An \eqn{n \times d} matrix \eqn{X} which factorizes
\eqn{M} as \eqn{M = XX^{\intercal}}. If \eqn{d=0}, \code{ase} returns
an \eqn{n \times 1} matrix of zeros.
}
\description{
\code{ase} calculates the \eqn{d}-dimensional adjacency spectral embedding of a symmetric,
positive semi-definite \eqn{n \times n} matrix \eqn{M}.
}
