% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\title{Inverse logistic link function}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
\code{expit} applies the inverse logistic link function \eqn{f(x) = e^x / (1+e^x)}.
}
