% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Logistic link function}
\usage{
logit(x,tol=1e-6)
}
\arguments{
\item{x}{A numeric vector with values in the interval [0,1].}

\item{tol}{A positive scalar which bounds the entries of x away from 0 and 1
for numerical stability.
Defaults to \code{tol=1e-6}}
}
\description{
\code{logit} applies the logistic link function \eqn{f(x) = log(x / (1-x))}.
}
